requirements.txt
asgiref==3.2.10
certifi==2020.6.20
Django==3.0.8
django-ckeditor==5.9.0
django-haystack==2.8.1
django-js-asset==1.2.2
django-model-utils==4.0.0
django-mptt==0.11.0
django-notifications-hq==1.6.0
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
elasticsearch==2.4.1
importlib-metadata==1.7.0
itsdangerous==1.1.0
jsonfield==3.1.0
Markdown==3.2.2
Pillow==7.2.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
swapper==1.1.2.post1
urllib3==1.25.9
zipp==3.1.0
